Add data files via upload
"F1 participants.txt" is the text file with the data for each participant.
"F1 simulator.txt" is the text file that my friend sent me as a starting point for me to understand how the simulation should end up. I have edited the text to make it more readable, but some parts are still in Norwegian as of now. One of the participants, Vettel, has wrong total points before the first race round.